+++ app/home/components/filters/FilterItem.tsx
@@ -1,5 +1,5 @@
-import { Checkbox } from '~/components/ui/checkbox';
-import { Label } from '~/components/ui/label';
+import { Checkbox } from '~/components/ui/Checkbox';
+import { Label } from '~/components/ui/Label';
 
 interface FilterItemProps {
   metric: string;
