+++ tsconfig.json
@@ -9,6 +9,7 @@
     "jsx": "react-jsx",
     "rootDirs": [".", "./.react-router/types"],
     "baseUrl": ".",
+    "forceConsistentCasingInFileNames": true,
     "paths": {
       "~/*": ["./app/*"]
     },
